adminUsers con alertas funcionando

+++ src/app/components/adminUsers/form-find-user/form-find-user.component.ts
@@ -5,7 +5,7 @@ import { FormBuilder, FormControl, FormGroup, ReactiveFormsModule, Validators } 
 import { CommonModule } from '@angular/common';
 import { Router } from '@angular/router';
 import { debounceTime, distinctUntilChanged, map } from 'rxjs';
-
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-form-find-user',
@@ -15,19 +15,19 @@ import { debounceTime, distinctUntilChanged, map } from 'rxjs';
   styleUrl: './form-find-user.component.scss'
 })
 export class FormFindUserComponent implements OnInit {
-  users: User[]=[]
-  usersFiltered: User [] =[]
-  searchControl= new FormControl('')
+  users: User[] = [];
+  usersFiltered: User [] = [];
+  searchControl= new FormControl('');
 
   searcherUser: FormGroup = this.FormBuilder.group({
-  user: new FormControl(null, [Validators.required])
-  })
+    user: new FormControl(null, [Validators.required])
+  });
 
-  constructor (
+  constructor(
     private UserService: UserService,
     private FormBuilder: FormBuilder, 
-    private route:Router
-  ){}
+    private route: Router
+  ) {}
 
   ngOnInit(): void {
     this.fetchAllUsers()
@@ -37,12 +37,12 @@ export class FormFindUserComponent implements OnInit {
   //Cargar todos los usuarios
   fetchAllUsers(): void {
     this.UserService.getAllUsers().subscribe({
-      next:(res:any)=>{
-        this.users=res as User[],
-        this.usersFiltered =res
-      }, 
-      error: (err:any)=>console.log ('Error al cargar usuarios', err)
-    })
+      next: (res:any) => {
+        this.users = res as User[];
+        this.usersFiltered = res;
+      },
+      error: (err: any) => console.log ('Error al cargar usuarios', err)
+    });
   }
 
   //escucha de cambios al escribir en la barra de busqueda
@@ -50,41 +50,61 @@ export class FormFindUserComponent implements OnInit {
     this.searchControl.valueChanges.pipe(
       debounceTime(300),
       distinctUntilChanged(),
-      map((value)=> value||'')
-    )
-    .subscribe((term)=>{
+      map((value) => value || '')
+    ).subscribe((term) => {
       this.filterUsers(term as string)
-    })
+    });
   }
 
-  filterUsers(searchTerm:string):void{
-    if (!searchTerm){
-      this.usersFiltered = this.users
-    }else{
-      const lowerTerm = searchTerm.toLowerCase()
+  filterUsers(searchTerm: string): void {
+    if (!searchTerm) {
+      this.usersFiltered = this.users;
+    } else {
+      const lowerTerm = searchTerm.toLowerCase();
       this.usersFiltered = this.users.filter(
-        (user)=> 
+        (user) => 
           user.name.toLowerCase().includes(lowerTerm) || 
           user.surname.toLowerCase().includes(lowerTerm) ||
           user.license.toLowerCase().includes(lowerTerm) ||
           user.email.toLowerCase().includes(lowerTerm)
-      ) 
+      );
     }
   }
 
-  navigateToEditUser(userId:any):void{
-    this.route.navigate([`/adminUsers/editUser/${userId}`])
+  navigateToEditUser(userId: any):void {
+    this.route.navigate([`/adminUsers/editUser/${userId}`]);
   }
 
-  deleteUser(userId:any): void{
-    this.UserService.deleteUser(userId).subscribe({
-      next:(res)=>{
-        this.fetchAllUsers()
-      },
-      error:(err) =>
-        console.error ('Error borrado usuasrio:', err)
-      })
-    }
-
-
+  deleteUser(userId: any): void {
+    Swal.fire({
+      title: '¿Estás seguro?',
+      text: '¡No podrás recuperar este usuario!',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: '¡Sí, eliminalo!',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.UserService.deleteUser(userId).subscribe({
+          next: (res) => {
+            this.fetchAllUsers();
+            Swal.fire(
+              '¡Eliminado!',
+              '¡El usuario ha sido eliminado!',
+              'success'
+            );
+          },
+          error: (err) => {
+            console.error ('Error eliminando el usuario: ', err);
+            Swal.fire(
+              '¡Error!',
+              'Error al borrar el usuario.',
+              'error'
+            );
+          },
+        });
+      }
+    });
+  }
 }
